+++ TDEngine2/include/core/CProjectSettings.h
@@ -79,7 +79,7 @@ namespace TDEngine2
 #endif
 
 				std::string mDefaultSkyboxMaterial = "DefaultResources/Materials/DefaultSkybox.material";
-				std::string mVolumetricCloudsMainShader = "Shaders/Default/VolumetricClouds.cshader";
+				std::string mVolumetricCloudsMainShader = "Shaders/Default/Volumetrics/VolumetricClouds.cshader";
 
 				std::string mShaderCachePathPattern = "ProjectResources/ShadersPack_{0}.cache";
 				bool        mIsShaderCacheEnabled = true;